Create 3-hier-fp.tcl
This is continuation of Posts 1 and 2 which adress the block level synthesis and top level synthesis of the opensparct2 processor . This post concentrates on building the floorplan/SDC etc suitable for block level P&R implementation. Due to short of time and bandwidth, I couldnt get a chance to build the floorplan for full chip “cpu” module, but to keep the momentum going and for enthusiasts, I built the flow for the “rtx” module which is quite big ( around 7M+ instances ) below and can act as good testcase to build the reference flow. Anyone who has the full cpu fix-time-final volcano from your top level synthesis run, can load it in place rtx or alternatively import the full chip netlist and follow the rest of the flow.

Folks can re-use the P&R implementation script for block level  from my Nova Processor Core RTL-GDSII post .

We are using channel style and the script below basically does the following :

 

1. Floorplan creation
2. Domain creation
3. Partitions the design based on the configs
4. Shapes the floorplan
5. Pin Assignment based on Global Routes
6. Power Ring/Mesh/pin tapping/via creation
7. Buffer long wires
8. Hierarchical top level CTS
9. Block level budget creation
10. Block level P&R implementation of the partitions created above

Exploration using Blackboxes is also covered.